Seperate 3D functionality and helper functions
common code for both projections is now in `helperfunctions.py`
3D plotting code `plotfunctions3D.py`
1D/2D plotting code `plotfunctions.py`

Refs #21720
