+++ Framework/PythonInterface/mantid/plots/CMakeLists.txt
@@ -1,6 +1,8 @@
 set ( PY_FILES
   __init__.py
-  functions.py
+  helperfunctions.py
+  plotfunctions.py
+  plotfunctions3D.py
 )
