+++ Framework/PythonInterface/mantid/plots/plotfunctions.py
@@ -19,65 +19,13 @@ from __future__ import (absolute_import, division, print_function)
 import numpy
 import mantid.kernel
 import mantid.api
-import mantid.dataobjects
+import mantid.plots.helperfunctions as common
 import matplotlib.colors
 
-from mantid.dataobjects import EventWorkspace, Workspace2D, MDHistoWorkspace
 
-
-def validate_args(*args):
-    return len(args) > 0 and (isinstance(args[0], EventWorkspace) or
-                              isinstance(args[0], Workspace2D) or
-                              isinstance(args[0], MDHistoWorkspace))
-
-
-def _getDistribution(workspace, **kwargs):
-    '''Determine whether or not the data is a distribution. The value in
-    the kwargs wins. Applies to Matrix workspaces only
-    :param workspace: :class:`mantid.api.MatrixWorkspace` to extract the data from'''
-    distribution = kwargs.pop('distribution', workspace.isDistribution())
-    return bool(distribution), kwargs
-
-
-def _getNormalization(mdworkspace, **kwargs):
-    '''gets the normalization flag of an MDHistoWorkspace. For workspaces
-    derived similar to MSlice/Horace, one needs to average data, the so-called
-    "number of events" normalization.
-    :param mdworkspace: :class:`mantid.api.IMDHistoWorkspace` to extract the data from'''
-    normalization = kwargs.pop('normalization', mdworkspace.displayNormalizationHisto())
-    return normalization, kwargs
-
-
-def getAxesLabels(workspace):
-    ''' get axis labels from a Workspace2D or an MDHistoWorkspace
-    Returns a tuple. The first element is the quantity label, such as "Intensity" or "Counts".
-    All other elements in the tuple are labels for axes.
-    Some of them are latex formatted already.
-    :param workspace: :class:`mantid.api.MatrixWorkspace` or :class:`mantid.api.IMDHistoWorkspace`
-    '''
-    if isinstance(workspace,mantid.dataobjects.MDHistoWorkspace):
-        axes = ['Intensity']
-        dims = workspace.getNonIntegratedDimensions()
-        for d in dims:
-            axis_title = d.name.replace('DeltaE', '$\Delta E$')
-            axis_unit = d.getUnits().replace('Angstrom^-1', '$\AA^{-1}$')
-            axis_unit = axis_unit.replace('DeltaE', 'meV')
-            axis_unit = axis_unit.replace('Angstrom', '$\AA$')
-            axis_unit = axis_unit.replace('MomentumTransfer', '$\AA^{-1}$')
-            axes.append('{0} ({1})'.format(axis_title, axis_unit))
-    else:
-        '''For matrix workspaces, return a tuple of ``(YUnit, <other units>)``'''
-        axes = [workspace.YUnit()]  # TODO: deal with distribution
-        for index in range(workspace.axes()):
-            axis = workspace.getAxis(index)
-            unit = axis.getUnit()
-            if len(str(unit.symbol())) > 0:
-                unit = '{} (${}$)'.format(unit.caption(), unit.symbol().latex())
-            else:
-                unit = unit.caption()
-            axes.append(unit)
-    return tuple(axes)
-
+# ================================================
+# Helper functions
+# ================================================
 
 def _getWkspIndexDistAndLabel(workspace, **kwargs):
     '''
@@ -115,7 +63,7 @@ def _getWkspIndexDistAndLabel(workspace, **kwargs):
         else:
             kwargs['label'] = 'spec {0}'.format(specNum)
 
-    (distribution, kwargs) = _getDistribution(workspace, **kwargs)
+    (distribution, kwargs) = common.get_distribution(workspace, **kwargs)
 
     return wkspIndex, distribution, kwargs
 
@@ -151,7 +99,7 @@ def _getSpectrum(workspace, wkspIndex, distribution, withDy = False, withDx = Fa
             y = y / (x[1:] - x[0:-1])
             if dy is not None:
                 dy = dy / (x[1:] - x[0:-1])
-        x = points_from_boundaries(x)
+        x = common.points_from_boundaries(x)
     y = numpy.ma.masked_invalid(y)
     if dy is not None:
         dy = numpy.ma.masked_invalid(dy)
@@ -195,7 +143,7 @@ def _getMatrix2DData(workspace, distribution, histogram2D=False):
         else:
             x = .5*(x[:, 0:-1]+x[:, 1:])
             if len(y) == z.shape[0]+1:
-                y = points_from_boundaries(y)
+                y = common.points_from_boundaries(y)
     else:
         if histogram2D:
             if _commonX(x):
@@ -207,7 +155,7 @@ def _getMatrix2DData(workspace, distribution, histogram2D=False):
                 y = boundaries_from_points(y)
         else:
             if len(y) == z.shape[0]+1:
-                y = points_from_boundaries(y)
+                y = common.points_from_boundaries(y)
     y = numpy.tile(y, x.shape[1]).reshape(x.shape[1], x.shape[0]).transpose()
     z = numpy.ma.masked_invalid(z)
     return x, y, z
@@ -264,18 +212,6 @@ def _getUnevenData(workspace, distribution):
     return x, y, z
 
 
-def _dim2array(d):
-    '''
-    Create a numpy array containing bin centers along the dimension d
-    :param d: an :class:`mantid.geometry.IMDDimension` object
-
-    returns: bin boundaries for dimension d
-    '''
-    dmin = d.getMinimum()
-    dmax = d.getMaximum()
-    return numpy.linspace(dmin, dmax, d.getNBins()+1)
-
-
 def boundaries_from_points(input_array):
     '''
     The function tries to guess bin boundaries from bin centers
@@ -291,90 +227,23 @@ def boundaries_from_points(input_array):
                              [(3*input_array[-1]-input_array[-2])*0.5]))
 
 
-def points_from_boundaries(input_array):
-    '''
-    The function returns bin centers from bin boundaries
-    :param input_array: a :class:`numpy.ndarray` of bin boundaries
-    '''
-    assert isinstance(input_array, numpy.ndarray), 'Not a numpy array'
-    if len(input_array) < 2:
-        raise ValueError('could not get centers from less than two boundaries')
-    return .5*(input_array[0:-1]+input_array[1:])
-
-
-def _getMDData(workspace, normalization, withError=False):
-    '''
-    generic function to extract data from an MDHisto workspace
-    :param workspace: :class:`mantid.api.IMDHistoWorkspace` containing data
-    :param normalization: if :class:`mantid.api.MDNormalization.NumEventsNormalization`
-        it will divide intensity by the number of corresponding MDEvents
-    returns a tuple containing bin boundaries for each dimension, the (maybe normalized)
-    signal and error arrays
-    '''
-    dims = workspace.getNonIntegratedDimensions()
-    dimarrays = [_dim2array(d) for d in dims]
-    # get data
-    data = workspace.getSignalArray()*1.
-    if normalization == mantid.api.MDNormalization.NumEventsNormalization:
-        nev = workspace.getNumEventsArray()
-        data /= nev
-    err = None
-    if withError:
-        err2 = workspace.getErrorSquaredArray()*1.
-        if normalization == mantid.api.MDNormalization.NumEventsNormalization:
-            err2 /= (nev*nev)
-        err = numpy.sqrt(err2)
-    data = data.squeeze().T
-    data = numpy.ma.masked_invalid(data)
-    if err is not None:
-        err = err.squeeze().T
-        err = numpy.ma.masked_invalid(err)
-    return dimarrays, data, err
-
-
 def _getMDData1D(workspace, normalization):
     '''
     Function to transform data in an MDHisto workspace with exactly
     one non-integrated dimension into arrays of bin centers, data,
     and error, to be used in 1D plots (plot, scatter, errorbar)
     '''
-    coordinate, data, err = _getMDData(workspace, normalization, withError=True)
+    coordinate, data, err = common.get_md_data(workspace, normalization, withError=True)
     assert len(coordinate) == 1, 'The workspace is not 1D'
-    coordinate = points_from_boundaries(coordinate[0])
+    coordinate = common.points_from_boundaries(coordinate[0])
     return coordinate, data, err
-
-
-def _getMDData2D_bin_bounds(workspace, normalization):
-    '''
-    Function to transform data in an MDHisto workspace with exactly
-    two non-integrated dimension into arrays of bin boundaries in each
-    dimension, and data. To be used in 2D plots (pcolor, pcolorfast, pcolormesh)
-    Note return coordinates are 1d vectors. Use numpy.meshgrid to generate 2d versions
-    '''
-    coordinate, data, _ = _getMDData(workspace, normalization, withError=False)
-    assert len(coordinate) == 2, 'The workspace is not 2D'
-    return coordinate[0], coordinate[1], data
-
-
-def _getMDData2D_bin_centers(workspace,normalization):
-    '''
-    Function to transform data in an MDHisto workspace with exactly
-    two non-integrated dimension into arrays of bin centers in each
-    dimension, and data. To be used in 2D plots (contour, contourf,
-    tricontour, tricontourf, tripcolor)
-    Note return coordinates are 1d vectors. Use numpy.meshgrid to generate 2d versions
-    '''
-    x, y, data = _getMDData2D_bin_bounds(workspace, normalization)
-    x = points_from_boundaries(x)
-    y = points_from_boundaries(y)
-    return x, y, data
 
 
 def _setLabels1D(axes, workspace):
     '''
     helper function to automatically set axes labels for 1D plots
     '''
-    labels = getAxesLabels(workspace)
+    labels = common.get_axes_labels(workspace)
     axes.set_xlabel(labels[1])
     axes.set_ylabel(labels[0])
 
@@ -383,20 +252,14 @@ def _setLabels2D(axes, workspace):
     '''
     helper function to automatically set axes labels for 2D plots
     '''
-    labels = getAxesLabels(workspace)
+    labels = common.get_axes_labels(workspace)
     axes.set_xlabel(labels[1])
     axes.set_ylabel(labels[2])
 
 
-def _setLabels3D(axes, workspace):
-    '''
-    heler function to automatically set axis labels for 3D plots
-    '''
-    labels = getAxesLabels(workspace)
-    axes.set_xlabel(labels[1])
-    axes.set_ylabel(labels[2])
-    axes.set_zlabel(labels[0])
-
+# ========================================================
+# Plot functions
+# ========================================================
 
 def plot(axes, workspace, *args, **kwargs):
     '''
@@ -423,7 +286,7 @@ def plot(axes, workspace, *args, **kwargs):
     dimension
     '''
     if isinstance(workspace, mantid.dataobjects.MDHistoWorkspace):
-        (normalization, kwargs)=_getNormalization(workspace, **kwargs)
+        (normalization, kwargs) = common.get_normalization(workspace, **kwargs)
         (x, y, _) = _getMDData1D(workspace, normalization)
     else:
         (wkspIndex, distribution, kwargs) = _getWkspIndexDistAndLabel(workspace, **kwargs)
@@ -457,8 +320,8 @@ def errorbar(axes, workspace, *args, **kwargs):
     dimension
     '''
     if isinstance(workspace, mantid.dataobjects.MDHistoWorkspace):
-        (normalization, kwargs) = _getNormalization(workspace, **kwargs)
-        (x, y, dy)=_getMDData1D(workspace, normalization)
+        (normalization, kwargs) = common.get_normalization(workspace, **kwargs)
+        (x, y, dy) = _getMDData1D(workspace, normalization)
         dx = None
     else:
         (wkspIndex, distribution, kwargs) = _getWkspIndexDistAndLabel(workspace, **kwargs)
@@ -492,7 +355,7 @@ def scatter(axes, workspace, *args, **kwargs):
     dimension
     '''
     if isinstance(workspace, mantid.dataobjects.MDHistoWorkspace):
-        (normalization, kwargs) = _getNormalization(workspace, **kwargs)
+        (normalization, kwargs) = common.get_normalization(workspace, **kwargs)
         (x, y, _) = _getMDData1D(workspace, normalization)
     else:
         (wkspIndex, distribution, kwargs) = _getWkspIndexDistAndLabel(workspace, **kwargs)
@@ -518,10 +381,10 @@ def contour(axes, workspace, *args, **kwargs):
                           the normalization is mantid.api.MDNormalization.NumEventsNormalization
     '''
     if isinstance(workspace, mantid.dataobjects.MDHistoWorkspace):
-        (normalization, kwargs) = _getNormalization(workspace, **kwargs)
-        x, y, z = _getMDData2D_bin_centers(workspace, normalization)
+        (normalization, kwargs) = common.get_normalization(workspace, **kwargs)
+        x, y, z = common.get_md_data2d_bin_centers(workspace, normalization)
     else:
-        (distribution, kwargs) = _getDistribution(workspace, **kwargs)
+        (distribution, kwargs) = common.get_distribution(workspace, **kwargs)
         (x, y, z) = _getMatrix2DData(workspace, distribution, histogram2D=False)
     _setLabels2D(axes, workspace)
     return axes.contour(x, y, z, *args, **kwargs)
@@ -544,10 +407,10 @@ def contourf(axes, workspace, *args, **kwargs):
                           the normalization is mantid.api.MDNormalization.NumEventsNormalization
     '''
     if isinstance(workspace, mantid.dataobjects.MDHistoWorkspace):
-        (normalization, kwargs) = _getNormalization(workspace, **kwargs)
-        x, y, z = _getMDData2D_bin_centers(workspace, normalization)
+        (normalization, kwargs) = common.get_normalization(workspace, **kwargs)
+        x, y, z = common.get_md_data2d_bin_centers(workspace, normalization)
     else:
-        (distribution, kwargs) = _getDistribution(workspace, **kwargs)
+        (distribution, kwargs) = common.get_distribution(workspace, **kwargs)
         (x, y, z) = _getMatrix2DData(workspace, distribution, histogram2D=False)
     _setLabels2D(axes, workspace)
     return axes.contourf(x, y, z, *args, **kwargs)
@@ -612,11 +475,11 @@ def pcolor(axes, workspace, *args, **kwargs):
     '''
     _setLabels2D(axes, workspace)
     if isinstance(workspace, mantid.dataobjects.MDHistoWorkspace):
-        (normalization, kwargs) = _getNormalization(workspace, **kwargs)
-        x, y, z = _getMDData2D_bin_bounds(workspace, normalization)
+        (normalization, kwargs) = common.get_normalization(workspace, **kwargs)
+        x, y, z = common.get_md_data2d_bin_bounds(workspace, normalization)
     else:
         (aligned, kwargs) = _getDataUnevenFlag(workspace, **kwargs)
-        (distribution, kwargs) = _getDistribution(workspace, **kwargs)
+        (distribution, kwargs) = common.get_distribution(workspace, **kwargs)
         if aligned:
             kwargs['pcolortype'] = ''
             return _pcolorpieces(axes, workspace, distribution, *args, **kwargs)
@@ -643,11 +506,11 @@ def pcolorfast(axes, workspace, *args, **kwargs):
     '''
     _setLabels2D(axes, workspace)
     if isinstance(workspace, mantid.dataobjects.MDHistoWorkspace):
-        (normalization, kwargs) = _getNormalization(workspace, **kwargs)
-        x, y, z = _getMDData2D_bin_bounds(workspace, normalization)
+        (normalization, kwargs) = common.get_normalization(workspace, **kwargs)
+        x, y, z = common.get_md_data2d_bin_bounds(workspace, normalization)
     else:
         (aligned, kwargs) = _getDataUnevenFlag(workspace, **kwargs)
-        (distribution, kwargs) = _getDistribution(workspace, **kwargs)
+        (distribution, kwargs) = common.get_distribution(workspace, **kwargs)
         if aligned:
             kwargs['pcolortype'] = 'fast'
             return _pcolorpieces(axes, workspace, distribution, *args, **kwargs)
@@ -674,11 +537,11 @@ def pcolormesh(axes, workspace, *args, **kwargs):
     '''
     _setLabels2D(axes, workspace)
     if isinstance(workspace, mantid.dataobjects.MDHistoWorkspace):
-        (normalization, kwargs) = _getNormalization(workspace, **kwargs)
-        x, y, z = _getMDData2D_bin_bounds(workspace, normalization)
+        (normalization, kwargs) = common.get_normalization(workspace, **kwargs)
+        x, y, z = common.get_md_data2d_bin_bounds(workspace, normalization)
     else:
         (aligned, kwargs) = _getDataUnevenFlag(workspace, **kwargs)
-        (distribution, kwargs) = _getDistribution(workspace, **kwargs)
+        (distribution, kwargs) = common.get_distribution(workspace, **kwargs)
         if aligned:
             kwargs['pcolortype'] = 'mesh'
             return _pcolorpieces(axes, workspace, distribution, *args, **kwargs)
@@ -707,11 +570,11 @@ def tripcolor(axes, workspace, *args, **kwargs):
     See :meth:`matplotlib.axes.Axes.tripcolor` for more information.
     '''
     if isinstance(workspace, mantid.dataobjects.MDHistoWorkspace):
-        (normalization, kwargs) = _getNormalization(workspace, **kwargs)
-        x_temp, y_temp, z = _getMDData2D_bin_centers(workspace, normalization)
+        (normalization, kwargs) = common.get_normalization(workspace, **kwargs)
+        x_temp, y_temp, z = common.get_md_data2d_bin_centers(workspace, normalization)
         x, y = numpy.meshgrid(x_temp, y_temp)
     else:
-        (distribution, kwargs) = _getDistribution(workspace, **kwargs)
+        (distribution, kwargs) = common.get_distribution(workspace, **kwargs)
         (x, y, z) = _getMatrix2DData(workspace, distribution, histogram2D=False)
     _setLabels2D(axes, workspace)
 
@@ -738,11 +601,11 @@ def tricontour(axes, workspace, *args, **kwargs):
     See :meth:`matplotlib.axes.Axes.tricontour` for more information.
     '''
     if isinstance(workspace, mantid.dataobjects.MDHistoWorkspace):
-        (normalization, kwargs) = _getNormalization(workspace, **kwargs)
-        (x_temp, y_temp, z) = _getMDData2D_bin_centers(workspace, normalization)
+        (normalization, kwargs) = common.get_normalization(workspace, **kwargs)
+        (x_temp, y_temp, z) = common.get_md_data2d_bin_centers(workspace, normalization)
         (x, y) = numpy.meshgrid(x_temp, y_temp)
     else:
-        (distribution, kwargs) = _getDistribution(workspace, **kwargs)
+        (distribution, kwargs) = common.get_distribution(workspace, **kwargs)
         (x, y, z) = _getMatrix2DData(workspace, distribution, histogram2D=False)
     _setLabels2D(axes, workspace)
     # tricontour segfaults if many z values are not finite
@@ -777,11 +640,11 @@ def tricontourf(axes, workspace, *args, **kwargs):
     See :meth:`matplotlib.axes.Axes.tricontourf` for more information.
     '''
     if isinstance(workspace, mantid.dataobjects.MDHistoWorkspace):
-        (normalization, kwargs) = _getNormalization(workspace, **kwargs)
-        (x_temp, y_temp, z) = _getMDData2D_bin_centers(workspace, normalization)
+        (normalization, kwargs) = common.get_normalization(workspace, **kwargs)
+        (x_temp, y_temp, z) = common.get_md_data2d_bin_centers(workspace, normalization)
         (x, y) = numpy.meshgrid(x_temp, y_temp)
     else:
-        (distribution, kwargs) = _getDistribution(workspace, **kwargs)
+        (distribution, kwargs) = common.get_distribution(workspace, **kwargs)
         (x, y, z) = _getMatrix2DData(workspace, distribution, histogram2D=False)
     _setLabels2D(axes, workspace)
     # tricontourf segfaults if many z values are not finite
@@ -794,142 +657,3 @@ def tricontourf(axes, workspace, *args, **kwargs):
     y = y[condition]
     z = z[condition]
     return axes.tricontourf(x, y, z, *args, **kwargs)
-
-
-# =============================================================
-# 3D plotting functionality
-# =============================================================
-def _extract_3d_data(workspace, **kwargs):
-    if isinstance(workspace, mantid.dataobjects.MDHistoWorkspace):
-        normalization, kwargs = _getNormalization(workspace, **kwargs)
-        x_temp, y_temp, z = _getMDData2D_bin_centers(workspace, normalization)
-        x, y = numpy.meshgrid(x_temp, y_temp)
-    else:
-        distribution, kwargs = _getDistribution(workspace, **kwargs)
-        x, y, z = _getMatrix2DData(workspace, distribution, histogram2D=False)
-    return x, y, z
-
-
-def plot3D(axes, workspace, *args, **kwargs):
-    """
-    3D plots - line plots
-    :param axes: class:`matplotlib.axes.Axes3D` object that will do the plotting
-    :param workspace: :class:`mantid.api.MatrixWorkspace` or
-                      :class:`mantid.api.IMDHistoWorkspace` to extract the data from
-    :param zdir: Which direction to use as z ('x', 'y' or 'z') when plotting a 2D set.
-    """
-    x, y, z = _extract_3d_data(workspace, **kwargs)
-    _setLabels3D(axes, workspace)
-    return axes.plot(x, y, z, *args, **kwargs)
-
-
-def scatter(axes, workspace, *args, **kwargs):
-    """
-    Scatter plots
-    :param axes: class:`matplotlib.axes.Axes3D` object that will do the plotting
-    :param workspace: :class:`mantid.api.MatrixWorkspace` or
-                      :class:`mantid.api.IMDHistoWorkspace` to extract the data from
-    :param zdir:	Which direction to use as z ('x', 'y' or 'z') when plotting a 2D set.
-    :param s:	Size in points^2. It is a scalar or an array of the same length as x and y.
-    :param c:	A color. c can be a single color format string, or a sequence of color
-                specifications of length N, or a sequence of N numbers to be mapped to
-                colors using the cmap and norm specified via kwargs (see below). Note
-                that c should not be a single numeric RGB or RGBA sequence because that
-                is indistinguishable from an array of values to be colormapped.
-                c can be a 2-D array in which the rows are RGB or RGBA, however, including
-                the case of a single row to specify the same color for all points.
-    :param depthshade:	Whether or not to shade the scatter markers to give the appearance
-                        of depth. Default is True.
-    """
-    x, y, z = _extract_3d_data(workspace, **kwargs)
-    _setLabels3D(axes, workspace)
-    return axes.scatter(x, y, z, *args, **kwargs)
-
-
-def plot_wireframe(axes, workspace, *args, **kwargs):
-    """
-    Wire-frame plot
-    :param axes: class:`matplotlib.axes.Axes3D` object that will do the plotting
-    :param workspace: :class:`mantid.api.MatrixWorkspace` or
-                      :class:`mantid.api.IMDHistoWorkspace` to extract the data from
-    :param rstride: Array row stride (step size), defaults to 1
-    :param cstride: Array column stride (step size), defaults to 1
-    :param rcount:	Use at most this many rows, defaults to 50
-    :param ccount:	Use at most this many columns, defaults to 50
-    """
-    x, y, z = _extract_3d_data(workspace, **kwargs)
-    _setLabels3D(axes, workspace)
-    return axes.plot_wireframe(x, y, z, *args, **kwargs)
-
-
-def plot_surface(axes, workspace, *args, **kwargs):
-    """
-    Surface plots
-    :param axes: class:`matplotlib.axes.Axes3D` object that will do the plotting
-    :param workspace: :class:`mantid.api.MatrixWorkspace` or
-                      :class:`mantid.api.IMDHistoWorkspace` to extract the data from
-    :param rstride: Array row stride (step size), defaults to 1
-    :param cstride: Array column stride (step size), defaults to 1
-    :param rcount:	Use at most this many rows, defaults to 50
-    :param ccount:	Use at most this many columns, defaults to 50
-    :param color:	Color of the surface patches
-    :param cmap:	A colormap for the surface patches.
-    :param norm:	An instance of Normalize to map values to colors
-    :param vmin:	Minimum value to map
-    :param vmax:	Maximum value to map
-    :param shade:	Whether to shade the facecolors
-    :param facecolors:	Face colors for the individual patches
-    """
-    x, y, z = _extract_3d_data(workspace, **kwargs)
-    _setLabels3D(axes, workspace)
-    return axes.plot_surface(x, y, z, *args, **kwargs)
-
-
-def plot_trisurf(axes, workspace, *args, **kwargs):
-    """
-    Tri-Surface plots
-    :param axes: class:`matplotlib.axes.Axes3D` object that will do the plotting
-    :param workspace: :class:`mantid.api.MatrixWorkspace` or
-                      :class:`mantid.api.IMDHistoWorkspace` to extract the data from
-    :param color:	Color of the surface patches
-    :param cmap:	A colormap for the surface patches.
-    :param norm:	An instance of Normalize to map values to colors
-    :param vmin:	Minimum value to map
-    :param vmax:	Maximum value to map
-    :param shade:	Whether to shade the facecolors
-    """
-    x, y, z = _extract_3d_data(workspace, **kwargs)
-    _setLabels3D(axes, workspace)
-    return axes.plot_trisurf(x, y, z, *args, **kwargs)
-
-
-def contour(axes, workspace, *args, **kwargs):
-    """
-    Contour plots
-    :param axes: class:`matplotlib.axes.Axes3D` object that will do the plotting
-    :param workspace: :class:`mantid.api.MatrixWorkspace` or
-                      :class:`mantid.api.IMDHistoWorkspace` to extract the data from
-    :param extend3d:	Whether to extend contour in 3D (default: False)
-    :param stride:	Stride (step size) for extending contour
-    :param zdir:	The direction to use: x, y or z (default)
-    :param offset:	If specified plot a projection of the contour lines
-                    on this position in plane normal to zdir
-    """
-    x, y, z = _extract_3d_data(workspace, **kwargs)
-    _setLabels3D(axes, workspace)
-    return axes.contour(x, y, z, *args, **kwargs)
-
-
-def contourf(axes, workspace, *args, **kwargs):
-    """
-    Filled Contour plots
-    :param axes: class:`matplotlib.axes.Axes3D` object that will do the plotting
-    :param workspace: :class:`mantid.api.MatrixWorkspace` or
-                      :class:`mantid.api.IMDHistoWorkspace` to extract the data from
-    :param zdir:	The direction to use: x, y or z (default)
-    :param offset:	If specified plot a projection of the filled contour on this
-                    position in plane normal to zdir
-    """
-    x, y, z = _extract_3d_data(workspace, **kwargs)
-    _setLabels3D(axes, workspace)
-    return axes.contourf(x, y, z, *args, **kwargs)
