+++ Framework/PythonInterface/mantid/plots/plotfunctions3D.py
@@ -1,0 +1,168 @@
+#  This file is part of the mantid package
+#
+#  Copyright (C) 2017 mantidproject
+#
+#  This program is free software: you can redistribute it and/or modify
+#  it under the terms of the GNU General Public License as published by
+#  the Free Software Foundation, either version 3 of the License, or
+#  (at your option) any later version.
+#
+#  This program is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#   GNU General Public License for more details.
+#
+#  You should have received a copy of the GNU General Public License
+#  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+from __future__ import (absolute_import, division, print_function)
+
+import numpy
+
+import mantid.plots.helperfunctions as common
+import mantid.dataobjects
+
+
+def _set_labels_3d(axes, workspace):
+    """
+    Helper function to automatically set axis labels for 3D plots
+    """
+    labels = common.get_axes_labels(workspace)
+    axes.set_xlabel(labels[1])
+    axes.set_ylabel(labels[2])
+    axes.set_zlabel(labels[0])
+
+
+def _extract_3d_data(workspace, **kwargs):
+    if isinstance(workspace, mantid.dataobjects.MDHistoWorkspace):
+        normalization, kwargs = common.get_normalization(workspace, **kwargs)
+        x_temp, y_temp, z = common.getMDData2D_bin_centers(workspace, normalization)
+        x, y = numpy.meshgrid(x_temp, y_temp)
+    else:
+        distribution, kwargs = common.get_distribution(workspace, **kwargs)
+        x, y, z = common.getMatrix2DData(workspace, distribution, histogram2D=False)
+    return x, y, z
+
+
+def plot3d(axes, workspace, *args, **kwargs):
+    """
+    3D plots - line plots
+    :param axes: class:`matplotlib.axes.Axes3D` object that will do the plotting
+    :param workspace: :class:`mantid.api.MatrixWorkspace` or
+                      :class:`mantid.api.IMDHistoWorkspace` to extract the data from
+    :param zdir: Which direction to use as z ('x', 'y' or 'z') when plotting a 2D set.
+    """
+    x, y, z = _extract_3d_data(workspace, **kwargs)
+    _set_labels_3d(axes, workspace)
+    return axes.plot(x, y, z, *args, **kwargs)
+
+
+def scatter(axes, workspace, *args, **kwargs):
+    """
+    Scatter plots
+    :param axes: class:`matplotlib.axes.Axes3D` object that will do the plotting
+    :param workspace: :class:`mantid.api.MatrixWorkspace` or
+                      :class:`mantid.api.IMDHistoWorkspace` to extract the data from
+    :param zdir:	Which direction to use as z ('x', 'y' or 'z') when plotting a 2D set.
+    :param s:	Size in points^2. It is a scalar or an array of the same length as x and y.
+    :param c:	A color. c can be a single color format string, or a sequence of color
+                specifications of length N, or a sequence of N numbers to be mapped to
+                colors using the cmap and norm specified via kwargs (see below). Note
+                that c should not be a single numeric RGB or RGBA sequence because that
+                is indistinguishable from an array of values to be colormapped.
+                c can be a 2-D array in which the rows are RGB or RGBA, however, including
+                the case of a single row to specify the same color for all points.
+    :param depthshade:	Whether or not to shade the scatter markers to give the appearance
+                        of depth. Default is True.
+    """
+    x, y, z = _extract_3d_data(workspace, **kwargs)
+    _set_labels_3d(axes, workspace)
+    return axes.scatter(x, y, z, *args, **kwargs)
+
+
+def plot_wireframe(axes, workspace, *args, **kwargs):
+    """
+    Wire-frame plot
+    :param axes: class:`matplotlib.axes.Axes3D` object that will do the plotting
+    :param workspace: :class:`mantid.api.MatrixWorkspace` or
+                      :class:`mantid.api.IMDHistoWorkspace` to extract the data from
+    :param rstride: Array row stride (step size), defaults to 1
+    :param cstride: Array column stride (step size), defaults to 1
+    :param rcount:	Use at most this many rows, defaults to 50
+    :param ccount:	Use at most this many columns, defaults to 50
+    """
+    x, y, z = _extract_3d_data(workspace, **kwargs)
+    _set_labels_3d(axes, workspace)
+    return axes.plot_wireframe(x, y, z, *args, **kwargs)
+
+
+def plot_surface(axes, workspace, *args, **kwargs):
+    """
+    Surface plots
+    :param axes: class:`matplotlib.axes.Axes3D` object that will do the plotting
+    :param workspace: :class:`mantid.api.MatrixWorkspace` or
+                      :class:`mantid.api.IMDHistoWorkspace` to extract the data from
+    :param rstride: Array row stride (step size), defaults to 1
+    :param cstride: Array column stride (step size), defaults to 1
+    :param rcount:	Use at most this many rows, defaults to 50
+    :param ccount:	Use at most this many columns, defaults to 50
+    :param color:	Color of the surface patches
+    :param cmap:	A colormap for the surface patches.
+    :param norm:	An instance of Normalize to map values to colors
+    :param vmin:	Minimum value to map
+    :param vmax:	Maximum value to map
+    :param shade:	Whether to shade the facecolors
+    :param facecolors:	Face colors for the individual patches
+    """
+    x, y, z = _extract_3d_data(workspace, **kwargs)
+    _set_labels_3d(axes, workspace)
+    return axes.plot_surface(x, y, z, *args, **kwargs)
+
+
+def plot_trisurf(axes, workspace, *args, **kwargs):
+    """
+    Tri-Surface plots
+    :param axes: class:`matplotlib.axes.Axes3D` object that will do the plotting
+    :param workspace: :class:`mantid.api.MatrixWorkspace` or
+                      :class:`mantid.api.IMDHistoWorkspace` to extract the data from
+    :param color:	Color of the surface patches
+    :param cmap:	A colormap for the surface patches.
+    :param norm:	An instance of Normalize to map values to colors
+    :param vmin:	Minimum value to map
+    :param vmax:	Maximum value to map
+    :param shade:	Whether to shade the facecolors
+    """
+    x, y, z = _extract_3d_data(workspace, **kwargs)
+    _set_labels_3d(axes, workspace)
+    return axes.plot_trisurf(x, y, z, *args, **kwargs)
+
+
+def contour(axes, workspace, *args, **kwargs):
+    """
+    Contour plots
+    :param axes: class:`matplotlib.axes.Axes3D` object that will do the plotting
+    :param workspace: :class:`mantid.api.MatrixWorkspace` or
+                      :class:`mantid.api.IMDHistoWorkspace` to extract the data from
+    :param extend3d:	Whether to extend contour in 3D (default: False)
+    :param stride:	Stride (step size) for extending contour
+    :param zdir:	The direction to use: x, y or z (default)
+    :param offset:	If specified plot a projection of the contour lines
+                    on this position in plane normal to zdir
+    """
+    x, y, z = _extract_3d_data(workspace, **kwargs)
+    _set_labels_3d(axes, workspace)
+    return axes.contour(x, y, z, *args, **kwargs)
+
+
+def contourf(axes, workspace, *args, **kwargs):
+    """
+    Filled Contour plots
+    :param axes: class:`matplotlib.axes.Axes3D` object that will do the plotting
+    :param workspace: :class:`mantid.api.MatrixWorkspace` or
+                      :class:`mantid.api.IMDHistoWorkspace` to extract the data from
+    :param zdir:	The direction to use: x, y or z (default)
+    :param offset:	If specified plot a projection of the filled contour on this
+                    position in plane normal to zdir
+    """
+    x, y, z = _extract_3d_data(workspace, **kwargs)
+    _set_labels_3d(axes, workspace)
+    return axes.contourf(x, y, z, *args, **kwargs)
